+++ src/Solvberget.Web/app/views/login.html
@@ -5,4 +5,4 @@
     <input type="password" placeholder="Passord" ng-model="password" />
     <button type="button" class="btn btn-large" ng-click="login()">Logg inn</button>
 </div>
-<div ng-show="isLoggedIn">Du er allerede logget inn. <a ng-click="logout()" class="btn btn-small">Logg ut</a></div>
+<div ng-show="isLoggedIn">Du er logget inn. <a ng-click="logout()" class="btn btn-small">Logg ut</a></div>
