+++ test/Solvberget.Nancy.Tests/packages.config
@@ -4,6 +4,7 @@
   <package id="CsQuery" version="1.3.4" targetFramework="net45" />
   <package id="FakeItEasy" version="1.14.0" targetFramework="net45" />
   <package id="Nancy" version="0.21.1" targetFramework="net45" />
+  <package id="Nancy.Authentication.Stateless" version="0.21.1" targetFramework="net45" />
   <package id="Nancy.Bootstrappers.Autofac" version="0.21.1" targetFramework="net45" />
   <package id="Nancy.Testing" version="0.21.1" targetFramework="net45" />
   <package id="Should" version="1.1.20" targetFramework="net45" />
